Change to config file check and trivial changes
Changed the if statement concerning the configuration file to readable
file check instead of normal file check because running the source
command on a non-readable file gives nasty errors. -r do however work on
directories too but the source command is silent about it so that's ok.

Made trivial changes to all the function lines. Looks better.

Bumped the version number a little bit.