Default audio files related bug fix
A bug made it possible to run Tim even though one of the needed audio
files was missing. This only happened when using a default setting for
either TIMER_CMD, WORK_CMD, BREAK_CMD or POMODORO_CMD + one or more
needed files was missing. The file check never happened.

I also added better debug output so if a setting needs a default audio
file that can't be found it's easier to spot it.